Extract SampleTreeNode into it's own file
